pyproject.toml
[project]
name = "python-whatsapp-bot"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "flask>=3.1.0",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
